pyproject.toml
[tool.poetry]
name = "elements_assignment"
version = "0.1.0"
description = ""
authors = ["thougtgeek <contact@thoughtgeek.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0.5"
djangorestframework = "^3.13.1"
uWSGI = "^2.0.20"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
django-csvimport = "^2.16"
django-cacheops = "^6.1"
coverage = "^6.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
